Moved typescript definition file to src/typings/
