+++ features/step_definitions/typescript_steps.ts
@@ -1,4 +1,4 @@
-import {defineSupportCode, StepDefinitions} from '../../dist/cucumber';
+import {defineSupportCode, StepDefinitions} from '../../src/typings/cucumber';
 
 defineSupportCode(function(stepDef:StepDefinitions){
     stepDef.Given("Given a typescript feature", () => {
